ref: restructure
