+++ elasticai/explorer/training/data.py
@@ -1,16 +1,11 @@
 from abc import abstractmethod
 from dataclasses import dataclass, field
 import logging
-import os
 from pathlib import Path
 from typing import Callable, List, Optional, Type, Union
-from iesude.data import DataSet
-from iesude.data.archives import PlainFile
 import numpy as np
-import owncloud
 import pandas as pd
 
-from iesude.data.extractable import ExtractableFn
 from torch.utils.data import Dataset
 from torchvision.datasets import MNIST
 from torchvision.transforms import Compose
@@ -97,34 +92,3 @@ class DatasetInfo:
     dataset_location: Path
     transform: Compose | None = None
     validation_split_ratio: List[float] = field(default_factory=lambda: [0.8, 0.2])
-
-
-def get_file_from_sciebo(
-    path_to_save: str,
-    file_path_in_sciebo: str,
-    file_type: ExtractableFn,
-):
-    if os.path.isfile(path_to_save) or (
-        os.path.isdir(path_to_save) and os.listdir(path_to_save)
-    ):
-        return
-
-    for i in range(10):
-        try:
-            if file_type is PlainFile:
-                dataset = DataSet(file_path=file_path_in_sciebo, file_type=file_type)
-                parent = Path(path_to_save).parent
-                dataset.download(parent)
-                save_path = Path(path_to_save).parent.parent / Path(file_path_in_sciebo)
-                os.renames(save_path, path_to_save)
-
-            else:
-                dataset = DataSet(file_path=file_path_in_sciebo, file_type=file_type)
-                dataset.download(path_to_save)
-        except owncloud.HTTPResponseError as err:
-            if i < 10:
-                continue
-            else:
-                raise err
-        else:
-            break
