+++ tests/system_tests/test_configs/deployment_config.yaml
@@ -2,10 +2,10 @@
 #If parameters are not specified inside the config, they are set to default values.
 DeploymentConfig:
   #The Hostname of target device.
-  target_name: "transpi4.local"
+  target_name: "transfair.local"
 
   #The Username/user-account of the target device to connect to.
-  target_user: "transfair"
+  target_user: "robin"
 
   target_platform_type: "rpi5"
   #The path (on host) to additional pre-compiled libraries neccessary for execution on target.
